+++ src/main/java/com/pengu/hammercore/utils/ColorNamePicker.java
@@ -225,7 +225,7 @@ public class ColorNamePicker
 		
 		public int computeMSE(int pixR, int pixG, int pixB)
 		{
-			return (int) (((pixR - r) * (pixR - r) + (pixG - g) * (pixG - g) + (pixB - b) * (pixB - b)) / 3);
+			return ((pixR - r) * (pixR - r) + (pixG - g) * (pixG - g) + (pixB - b) * (pixB - b)) / 3;
 		}
 		
 		public int getR()
